// apps/git-client/src/generated/ConflictContent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ConflictContent = { path: string, base: string | null, local: string | null, remote: string | null, result: string | null, binary: boolean, localLabel: string, remoteLabel: string, };